Review suggestions from @samuelbray32
